+++ src/Entity/BlogPost.php
@@ -7,6 +7,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Validator\Constraints as Assert;
 
 #[ORM\Entity(repositoryClass: BlogPostRepository::class)]
 #[ORM\Index(columns: ['published_at'], name: 'idx_blog_post_date')]
@@ -19,6 +20,7 @@ class BlogPost
     private ?int $id = null;
 
     #[ORM\Column(type: Types::TEXT)]
+    #[Assert\NotBlank]
     private string $content = '';
 
     #[ORM\Column(type: Types::DATETIME_IMMUTABLE)]